contract as type + unify contract+interface
This is my first attempt to unify contract and
interfaces.

I think it gives backward compatability, and unifies
contracts and interfaces into one thing.

The difference in boxing or not is expressed in how
contracts are used.  When applied to parameters,
no boxing.  When a contract is unary and is applied as
a type, boxing, like interfaces now.